+++ tests/trace/basic/ExampleTagTest.m
@@ -1,0 +1,27 @@
+% Taken from https://www.mathworks.com/help/matlab/matlab_prog/tag-unit-tests.html
+
+classdef ExampleTagTest < matlab.unittest.TestCase
+    methods (Test)
+        function testA (testCase)
+            % test code
+        end
+    end
+    methods (Test, TestTags = {'Unit'})
+        function testB (testCase)
+            % test code
+        end
+        function testC (testCase)
+            % test code
+        end
+    end
+    methods (Test, TestTags = {'Unit','FeatureA'})
+        function testD (testCase)
+            % test code
+        end
+    end
+    methods (Test, TestTags = {'System','FeatureA'})
+        function testE (testCase)
+            % test code
+        end
+    end
+end